requirements.txt
# backwards compatibility
pathlib2

# external requirements
Sphinx
python-dotenv>=0.5.1

webargs>=5.5.2
marshmallow

deepaas>=1.0.0
flaat>=0.5.1

Pillow == 6.1.0
fpdf==1.7.2
